Move Values classes that shouldn't be directly used into Helpers module
