+++ History.txt
@@ -1,6 +1,7 @@
 === Unreleased
  * Add changelog metadata to gemspec - Juri Hahn
  * Attempt to rescue timezone info when given a nonstandard tzid with no VTIMEZONE
+ * Move Values classes that shouldn't be directly used into Helpers module
 
 === 2.9.0 2023-08-11
  * Always include the VALUE of Event URLs for improved compatibility - Sean Kelley
