+++ lib/icalendar/values/date_time.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 require 'date'
-require_relative 'time_with_zone'
+require_relative 'helpers/time_with_zone'
 
 module Icalendar
   module Values
 
     class DateTime < Value
-      include TimeWithZone
+      include Helpers::TimeWithZone
 
       FORMAT = '%Y%m%dT%H%M%S'
