+++ spec/roundtrip_spec.rb
@@ -14,7 +14,7 @@ describe Icalendar do
       event = Icalendar::Event.new
       parsed = Icalendar::Calendar.parse(source).first
       event.rdate = parsed.events.first.rdate
-      expect(event.rdate.first).to be_kind_of Icalendar::Values::Array
+      expect(event.rdate.first).to be_kind_of Icalendar::Values::Helpers::Array
       expect(event.rdate.first.params_ical).to eq ";TZID=US-Mountain"
     end
   end
